Fix windows errors and restructure code to align with installation

+++ src/seabolt/src/bolt/mem.c
@@ -22,11 +22,11 @@
 #include "mem.h"
 #include "platform.h"
 
-void* BoltMem_reverse_copy(void* dest, const void* src, size_t n)
+void* BoltMem_reverse_copy(void* dest, const void* src, int64_t n)
 {
     char* dest_c = (char*) (dest);
     const char* src_c = (const char*) (src);
-    for (size_t i = 0; i<n; i++) {
+    for (int64_t i = 0; i<n; i++) {
         dest_c[i] = src_c[n-i-1];
     }
     return dest;
@@ -36,28 +36,28 @@ static int64_t __allocation = 0;
 static int64_t __peak_allocation = 0;
 static int64_t __allocation_events = 0;
 
-void* BoltMem_allocate(size_t new_size)
+void* BoltMem_allocate(int64_t new_size)
 {
     void* p = malloc(new_size);
-    int64_t new_allocation = BoltUtil_add(&__allocation, new_size);
-    int64_t peak_allocation = BoltUtil_add(&__peak_allocation, 0);
+    size_t new_allocation = BoltUtil_add(&__allocation, new_size);
+    size_t peak_allocation = BoltUtil_add(&__peak_allocation, 0);
     if (new_allocation>peak_allocation) BoltUtil_add(&__peak_allocation, new_allocation-peak_allocation);
     BoltUtil_increment(&__allocation_events);
     return p;
 }
 
-void* BoltMem_reallocate(void* ptr, size_t old_size, size_t new_size)
+void* BoltMem_reallocate(void* ptr, int64_t old_size, int64_t new_size)
 {
     void* p = realloc(ptr, new_size);
 
-    int64_t new_allocation = BoltUtil_add(&__allocation, -old_size+new_size);
-    int64_t peak_allocation = BoltUtil_add(&__peak_allocation, 0);
+    size_t new_allocation = BoltUtil_add(&__allocation, -old_size+new_size);
+    size_t peak_allocation = BoltUtil_add(&__peak_allocation, 0);
     if (__allocation>__peak_allocation) BoltUtil_add(&__peak_allocation, new_allocation-peak_allocation);
     BoltUtil_increment(&__allocation_events);
     return p;
 }
 
-void* BoltMem_deallocate(void* ptr, size_t old_size)
+void* BoltMem_deallocate(void* ptr, int64_t old_size)
 {
     if (ptr==NULL) {
         return NULL;
@@ -69,7 +69,7 @@ void* BoltMem_deallocate(void* ptr, size_t old_size)
     return NULL;
 }
 
-void* BoltMem_adjust(void* ptr, size_t old_size, size_t new_size)
+void* BoltMem_adjust(void* ptr, int64_t old_size, int64_t new_size)
 {
     if (new_size==old_size) {
         // In this case, the physical data storage requirement
@@ -103,7 +103,7 @@ void* BoltMem_adjust(void* ptr, size_t old_size, size_t new_size)
     return BoltMem_reallocate(ptr, old_size, new_size);
 }
 
-void* BoltMem_duplicate(const void* ptr, size_t ptr_size)
+void* BoltMem_duplicate(const void* ptr, int64_t ptr_size)
 {
     if (ptr==NULL) {
         return NULL;
